refactoring according to CR
